langgraph_impl/requirements.txt
# Core data
pandas>=2.0
numpy>=1.26
openpyxl>=3.1
python-dotenv>=1.0

# LangGraph ecosystem
langgraph>=0.2
langchain-openai>=0.1
langchain-core>=0.2

# Data models
pydantic>=2.0
typing_extensions>=4.0
